Split login managers into separate packages
