+++ radar-android-login-oauth2/build.gradle
@@ -1,0 +1,114 @@
+android {
+    compileSdkVersion 26
+    buildToolsVersion "26.0.1"
+
+    defaultConfig {
+        minSdkVersion 21
+        targetSdkVersion 26
+        versionCode versionCode
+        versionName version
+        manifestPlaceholders = ['appAuthRedirectScheme': 'org.radarcns.android']
+    }
+    lintOptions {
+        abortOnError false
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+
+            libraryVariants.all { variant ->
+                variant.outputs.each { output ->
+                    if (output.outputFile != null && output.outputFile.name.endsWith('.aar')) {
+                        output.outputFile = new File(output.outputFile.parent,
+                                "${archivesBaseName}-${android.defaultConfig.versionName}.aar")
+                    }
+                }
+            }
+        }
+    }
+}
+
+ext.moduleName = 'radar-android-login-oauth2'
+ext.description = 'RADAR Android OAuth2 LoginManager.'
+
+dependencies {
+    compile project(":")
+    compile 'net.openid:appauth:0.6.1'
+    compile 'io.jsonwebtoken:jjwt:0.7.0'
+}
+
+
+ext.sharedManifest = manifest {
+    attributes("Implementation-Title": moduleName,
+            "Implementation-Version": version)
+}
+
+task sourcesJar(type: Jar) {
+    classifier = 'sources'
+    from android.sourceSets.main.java.srcDirs
+    manifest.from sharedManifest
+}
+
+task javadoc(type: Javadoc) {
+    source = android.sourceSets.main.java.srcDirs
+    classpath += configurations.compile
+    classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
+}
+
+task javadocJar(type: Jar, dependsOn: javadoc) {
+    classifier = 'javadoc'
+    from javadoc.destinationDir
+    manifest.from sharedManifest
+}
+
+// add javadoc/source jar tasks as artifacts
+artifacts {
+    archives javadocJar
+    archives sourcesJar
+}
+
+install {
+    repositories.mavenInstaller {
+        pom.project {
+            groupId project.group
+            artifactId moduleName
+            version project.version
+        }
+        pom.withXml {
+            def root = asNode()
+            root.appendNode('description', description)
+            root.appendNode('name', moduleName)
+            root.appendNode('url', githubUrl)
+            root.children().last() + pomConfig
+        }
+    }
+}
+
+bintray {
+    user = project.hasProperty('bintrayUser') ? project.property('bintrayUser') : System.getenv('BINTRAY_USER')
+    key = project.hasProperty('bintrayApiKey') ? project.property('bintrayApiKey') : System.getenv('BINTRAY_API_KEY')
+    override = false
+    configurations = ['archives']
+    pkg {
+        repo = project.group
+        name = moduleName
+        userOrg = 'radar-cns'
+        desc = description
+        licenses = ['Apache-2.0']
+        websiteUrl = website
+        issueTrackerUrl = issueUrl
+        vcsUrl = githubUrl
+        githubRepo = githubRepoName
+        githubReleaseNotesFile = 'README.md'
+        version {
+            name = project.version
+            desc = description
+            vcsTag = System.getenv('TRAVIS_TAG')
+            released = new Date()
+            gpg {
+                sign = true //Determines whether to GPG sign the files. The default is false
+            }
+        }
+    }
+}
