+++ radar-android-login-oauth2/proguard-rules.pro
