+++ radar-android-login-oauth2/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest package="org.radarcns.android.auth.oauth2">
+
+    <application>
+
+    </application>
+
+</manifest>
