+++ radar-android-login-oauth2/src/main/java/org/radarcns/android/auth/oauth2/OAuth2StateManager.java
