+++ radar-android-login-oauth2/src/main/res/values/strings.xml
@@ -1,0 +1,2 @@
+<resources>
+</resources>
