+++ radar-android-login-qr/.gitignore
