+++ radar-android-login-qr/proguard-rules.pro
