+++ radar-android-login-qr/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<manifest package="org.radarcns.android.auth.qr">
+    <application>
+
+    </application>
+
+</manifest>
