--- radar-commons-android-oauth2/build.gradle
@@ -1,31 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.1"
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-        manifestPlaceholders = ['appAuthRedirectScheme': 'org.radarcns.android']
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-repositories {
-    jcenter()
-    maven { url 'https://oss.jfrog.org/artifactory/oss-snapshot-local' }
-}
-
-dependencies {
-    compile project(":")
-    compile 'net.openid:appauth:0.6.1'
-    compile 'io.jsonwebtoken:jjwt:0.7.0'
-}
