--- radar-commons-android-oauth2/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">radar-commons-android-oauth2</string>
-</resources>
