--- radar-commons-android-qr/build.gradle
@@ -1,29 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.1"
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-repositories {
-    jcenter()
-    maven { url 'https://oss.jfrog.org/artifactory/oss-snapshot-local' }
-}
-
-dependencies {
-    compile project(":")
-    compile 'com.journeyapps:zxing-android-embedded:3.5.0'
-}
